+++ ibmmq/mqistr.go
@@ -1,5 +1,3 @@
-// +build go1.13
-
 package ibmmq
 
 /*
@@ -781,6 +779,16 @@ func MQItoString(class string, value int) string {
 			s = ""
 		}
 
+	case "CEX":
+		switch v {
+		case -2:
+			s = "MQCEX_AS_PARENT"
+		case -1:
+			s = "MQCEX_NOLIMIT"
+		default:
+			s = ""
+		}
+
 	case "CFACCESS":
 		switch v {
 		case 0:
@@ -1425,6 +1433,8 @@ func MQItoString(class string, value int) string {
 			s = "MQCMDL_LEVEL_925"
 		case 930:
 			s = "MQCMDL_LEVEL_930"
+		case 931:
+			s = "MQCMDL_LEVEL_931"
 		default:
 			s = ""
 		}
@@ -2171,6 +2181,8 @@ func MQItoString(class string, value int) string {
 			s = "MQFIELD_WQR_CLWLQueueRank"
 		case 8015:
 			s = "MQFIELD_WQR_DefPutResponse"
+		case 8016:
+			s = "MQFIELD_WQR_CapExpiry"
 		default:
 			s = ""
 		}